dx1238: rename factory
